+++ src/shillelagh/adapters/api/gsheets/fields.py
@@ -2,14 +2,16 @@
 Custom fields for the GSheets adapter.
 """
 import datetime
+from distutils.util import strtobool
 from typing import Any
 from typing import List
 from typing import Optional
 from typing import Type
 
+from shillelagh.fields import External
 from shillelagh.fields import Field
+from shillelagh.fields import Internal
 from shillelagh.fields import Order
-from shillelagh.fields import StringBoolean
 from shillelagh.filters import Filter
 
 
@@ -18,53 +20,53 @@ from shillelagh.filters import Filter
 # sheet.
 DATETIME_CELL_FORMAT = "%m/%d/%Y %H:%M:%S"
 DATE_CELL_FORMAT = "%m/%d/%Y"
-TO_TIME_CELL_FORMAT = "%-I:%M:%S %p"
-FROM_TIME_CELL_FORMAT = "%I:%M:%S %p"
+TIME_CELL_FORMAT = "%I:%M:%S %p"
 
 # timestamp format used in SQL queries
 DATETIME_SQL_QUOTE = "%Y-%m-%d %H:%M:%S"
+DATE_SQL_QUOTE = "%Y-%m-%d"
 TIME_SQL_QUOTE = "%H:%M:%S"
 
 
-class GSheetsDateTime(Field[str, datetime.datetime]):
+def convert_pattern_to_format(pattern: Optional[str]) -> str:
     """
-    A GSheets timestamp.
+    Convert a Google Chart API pattern to a python time format.
 
-    The Google Chart API returns timestamps as a string in the following format:
+    Reference: https://developers.google.com/sheets/api/guides/formats?hl=en
+    """
+    # Google Chart API uses ICU to represent patterns. In Python we can use
+    # PyICU to parse the values given a pattern, but the library is difficult
+    # to install and big. For now we have hardcoded values.
+    formats = {
+        "M/d/yyyy H:mm:ss": DATETIME_CELL_FORMAT,
+        "M/d/yyyy": DATE_CELL_FORMAT,
+        "h:mm:ss am/pm": TIME_CELL_FORMAT,
+    }
 
-        Date(2020,0,1,0,0,0) => '2020-01-01 00:00:00'
+    if pattern not in formats:
+        raise NotImplementedError(
+            f'Unknown pattern "{pattern}". Please file a ticket at '
+            "https://github.com/betodealmeida/shillelagh/issues.",
+        )
 
-    (Note that the month is zero-indexed, probably to be compatible with
-    Javascript.)
+    return formats[pattern]
 
-    There are no timezones; instead, there is a global timezone for the whole
-    spreadsheet. The timezone can only be read if the user has set their
-    credentials, since the Google Sheets API used to read metadata about the
-    sheet requires authentication.
 
-    When the timezone is present and read succesfully all timestamps are
-    converted to it, both when fetching data as well as when inserting rows.
-
-    When inserting data timestamps should be formatted differently, and also
-    without a timezone:
-
-        '2020-01-01 00:00:00'
-
-    This field takes care of the conversion between `datetime.datetime` and the
-    two formats.
+class GSheetsField(Field[Internal, External]):
+    """
+    A base class for GSheets fields.
     """
 
-    type = "TIMESTAMP"
-    db_api_type = "DATETIME"
-
-    def __init__(
+    def __init__(  # pylint: disable=too-many-arguments
         self,
         filters: Optional[List[Type[Filter]]] = None,
         order: Order = Order.NONE,
         exact: bool = False,
+        pattern: Optional[str] = None,
         timezone: Optional[datetime.tzinfo] = None,
     ):
         super().__init__(filters, order, exact)
+        self.pattern = pattern
         self.timezone = timezone
 
     def __eq__(self, other: Any) -> bool:
@@ -75,127 +77,241 @@ class GSheetsDateTime(Field[str, datetime.datetime]):
             self.filters == other.filters
             and self.order == other.order
             and self.exact == other.exact
+            and self.pattern == other.pattern
             and self.timezone == other.timezone,
         )
 
+
+class GSheetsDateTime(GSheetsField[str, datetime.datetime]):
+    """
+    A GSheets timestamp.
+
+    The Google Chart API returns timestamps as a string encoded using an ICU
+    pattern. The default format is "M/d/yyyy H:mm:ss", and values look like
+    this:
+
+        "9/1/2018 0:00:00"
+
+    There are no timezones; instead, there is a global timezone for the whole
+    spreadsheet. The timezone can only be read if the user has set their
+    credentials, since the Google Sheets API used to read metadata about the
+    sheet requires authentication.
+
+    When the timezone is present and read succesfully all timestamps are
+    converted to it, both when fetching data as well as when inserting rows.
+    """
+
+    type = "TIMESTAMP"
+    db_api_type = "DATETIME"
+
     def parse(self, value: Optional[str]) -> Optional[datetime.datetime]:
-        if value is None:
+        # Google Charts API returns `None` for a NULL cell, while the Google
+        # Sheets API returns an empty string
+        if value is None or value == "":
             return None
 
-        args = [int(number) for number in value[len("Date(") : -1].split(",")]
-        args[1] += 1  # month is zero indexed in the response
-        return datetime.datetime(*args, tzinfo=self.timezone)  # type: ignore
+        format_ = convert_pattern_to_format(self.pattern)
+        timestamp = datetime.datetime.strptime(value, format_)
 
-    def format(self, value: Optional[datetime.datetime]) -> Optional[str]:
+        # Set the timestamp to the spreadsheet timezone, if any.
+        timestamp = timestamp.replace(tzinfo=self.timezone)
+
+        return timestamp
+
+    def format(self, value: Optional[datetime.datetime]) -> str:
+        # This method is used only when inserting or updating rows, so we
+        # encode NULLs as an empty string to match the Google Sheets API.
         if value is None:
-            return None
+            return ""
 
         # Google Sheets does not support timezones in datetime values, so we
         # convert all timestamps to the sheet timezone
         if self.timezone:
             value = value.astimezone(self.timezone)
 
-        return value.strftime(DATETIME_CELL_FORMAT) if value else None
+        format_ = convert_pattern_to_format(self.pattern)
+        return value.strftime(format_)
 
     def quote(self, value: Optional[str]) -> str:
-        if value is None:
-            return "NULL"
+        if value == "" or value is None:
+            return "null"
 
         # On SQL queries the timestamp should be prefix by "datetime"
-        value = datetime.datetime.strptime(value, DATETIME_CELL_FORMAT).strftime(
+        format_ = convert_pattern_to_format(self.pattern)
+        value = datetime.datetime.strptime(value, format_).strftime(
             DATETIME_SQL_QUOTE,
         )
         return f"datetime '{value}'"
 
 
-class GSheetsDate(Field[str, datetime.date]):
+class GSheetsDate(GSheetsField[str, datetime.date]):
     """
     A GSheets date.
 
-    The Google Chart API returns timestamps as a string in the following format:
+    The Google Chart API returns dates as a string encoded using an ICU
+    pattern. The default format is "M/d/yyyy", and values look like
+    this:
 
-        Date(2020,0,1) => '2020-01-01'
+        "9/1/2018"
 
-    (Note that the month is zero-indexed, probably to be compatible with
-    Javascript.)
-
-    When inserting data timestamps should be formatted differently:
-
-        '2020-01-01'
-
-    This field takes care of the conversion between `datetime.date` and the
-    two formats.
     """
 
     type = "DATE"
     db_api_type = "DATETIME"
 
     def parse(self, value: Optional[str]) -> Optional[datetime.date]:
-        """Parse a string like 'Date(2018,0,1)'."""
-        if value is None:
+        # Google Charts API returns `None` for a NULL cell, while the Google
+        # Sheets API returns an empty string
+        if value is None or value == "":
             return None
 
-        args = [int(number) for number in value[len("Date(") : -1].split(",")]
-        args[1] += 1  # month is zero indexed in the response (WTF, Google!?)
-        return datetime.date(*args)
+        format_ = convert_pattern_to_format(self.pattern)
+        return datetime.datetime.strptime(value, format_).date()
 
-    def format(self, value: Optional[datetime.date]) -> Optional[str]:
+    def format(self, value: Optional[datetime.date]) -> str:
         if value is None:
-            return None
-        return value.isoformat()
+            return ""
+
+        format_ = convert_pattern_to_format(self.pattern)
+        return value.strftime(format_)
 
     def quote(self, value: Optional[str]) -> str:
-        if value is None:
-            return "NULL"
+        if value == "" or value is None:
+            return "null"
 
         # On SQL queries the timestamp should be prefix by "date"
+        format_ = convert_pattern_to_format(self.pattern)
+        value = datetime.datetime.strptime(value, format_).strftime(DATE_SQL_QUOTE)
         return f"date '{value}'"
 
 
-class GSheetsTime(Field[List[int], datetime.time]):
+class GSheetsTime(GSheetsField[str, datetime.time]):
     """
     A GSheets time.
 
-    The Google Chart API return time objects as a list of numbers, corresponding
-    to hour, minute, and seconds.
+    The Google Chart API returns times as a string encoded using an ICU
+    pattern. The default format is "h:mm:ss am/pm", and values look like
+    this:
+
+        "5:00:00 PM"
+
     """
 
     type = "TIME"
     db_api_type = "DATETIME"
 
-    def parse(self, value: Optional[List[int]]) -> Optional[datetime.time]:
+    def parse(self, value: Optional[str]) -> Optional[datetime.time]:
         """
         Parse time of day as returned from the Google Chart API.
         """
-        if value is None:
+        # Google Charts API returns `None` for a NULL cell, while the Google
+        # Sheets API returns an empty string
+        if value is None or value == "":
             return None
 
-        return datetime.time(*value)  # type: ignore
+        format_ = convert_pattern_to_format(self.pattern)
+        return datetime.datetime.strptime(value, format_).time()
 
-    def format(self, value: Optional[datetime.time]) -> Optional[str]:  # type: ignore
+    def format(self, value: Optional[datetime.time]) -> str:
         if value is None:
-            return value
-        return value.strftime(TO_TIME_CELL_FORMAT)
+            return ""
 
-    def quote(self, value: Optional[str]) -> str:  # type: ignore
-        if value is None:
-            return "NULL"
+        format_ = convert_pattern_to_format(self.pattern)
+        return value.strftime(format_)
+
+    def quote(self, value: Optional[str]) -> str:
+        if value == "" or value is None:
+            return "null"
 
         # On SQL queries the timestamp should be prefix by "timeofday"
-        value = datetime.datetime.strptime(value, FROM_TIME_CELL_FORMAT).strftime(
+        format_ = convert_pattern_to_format(self.pattern)
+        value = datetime.datetime.strptime(value, format_).strftime(
             TIME_SQL_QUOTE,
         )
         return f"timeofday '{value}'"
 
 
-class GSheetsBoolean(StringBoolean):
+class GSheetsBoolean(GSheetsField[str, bool]):
     """
     A GSheets boolean.
 
-    Booleans in the Google Chart API are represented as lowercase strings.
+    Booleans in the Google Chart API are return as a string, either "TRUE"
+    of "FALSE".
     """
 
-    def quote(self, value: Optional[str]) -> str:
+    type = "BOOLEAN"
+    db_api_type = "NUMBER"
+
+    def parse(self, value: Optional[str]) -> Optional[bool]:
+        # Google Charts API returns `None` for a NULL cell, while the Google
+        # Sheets API returns an empty string
+        if value is None or value == "":
+            return None
+
+        return bool(strtobool(value))
+
+    def format(self, value: Optional[bool]) -> str:
         if value is None:
-            return "NULL"
+            return ""
+
+        return "TRUE" if value else "FALSE"
+
+    def quote(self, value: Optional[str]) -> str:
+        if value == "" or value is None:
+            return "null"
         return value.lower()
+
+
+class GSheetsNumber(GSheetsField[str, float]):
+    """
+    A GSheets number.
+
+    The Google Chart/Sheets APIs return "numbers" only, encoded as strings.
+    """
+
+    type = "REAL"
+    db_api_type = "NUMBER"
+
+    def parse(self, value: Optional[str]) -> Optional[float]:
+        if value is None or value == "":
+            return None
+
+        try:
+            return int(value)
+        except ValueError:
+            pass
+
+        return float(value)
+
+    def format(self, value: Optional[float]) -> str:
+        if value is None:
+            return ""
+
+        return str(value)
+
+    def quote(self, value: Optional[str]) -> str:
+        if value == "" or value is None:
+            return "null"
+
+        return str(value)
+
+
+class GSheetsString(GSheetsField[str, str]):
+    """
+    A GSheets string.
+    """
+
+    type = "TEXT"
+    db_api_type = "STRING"
+
+    def parse(self, value: Optional[str]) -> Optional[str]:
+        return None if value == "" else value
+
+    def format(self, value: Optional[str]) -> str:
+        return "" if value is None else value
+
+    def quote(self, value: Optional[str]) -> str:
+        if value == "" or value is None:
+            return "null"
+
+        return f"'{value}'"